fix: better mint flow
